Fix system_code using the wrong erased type on Windows

+++ include/system_code.hpp
@@ -40,7 +40,7 @@ For Windows, these might be:
     - `nt_code` (`LONG`)
     - `win32_code` (`DWORD`)
 
-So the erased type is `LONG` on Windows, as that can represent all of
+So the erased type is `intptr_t` on Windows, as that can represent all of
 the above.
 
 For POSIX, `posix_code` (`int`) is possible, so this erased type is `int`
@@ -48,7 +48,7 @@ i.e. this is the type alias on POSIX, not the above:
 
 `using system_code = status_code<erased<int>>`
 */
-using system_code = status_code<erased<long>>;
+using system_code = status_code<erased<intptr_t>>;
 SYSTEM_ERROR2_NAMESPACE_END
 
 #else
